Refactor some components of explorer
